Refactoring callback class into Player file
