+++ client/android/WorldScope/app/src/main/java/com/litmus/worldscope/LitmusPlayer.java
@@ -2,8 +2,10 @@ package com.litmus.worldscope;
 
 import android.content.Context;
 import android.media.MediaCodec;
+import android.media.MediaDrm;
 import android.os.Handler;
 import android.os.Looper;
+import android.text.TextUtils;
 import android.util.Log;
 import android.view.Surface;
 
@@ -21,7 +23,10 @@ import com.google.android.exoplayer.drm.StreamingDrmSessionManager;
 import com.google.android.exoplayer.upstream.BandwidthMeter;
 import com.google.android.exoplayer.upstream.DefaultBandwidthMeter;
 import com.google.android.exoplayer.util.PlayerControl;
+import com.google.android.exoplayer.util.Util;
 
+import java.io.IOException;
+import java.util.UUID;
 import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.concurrent.locks.Condition;
 import java.util.concurrent.locks.Lock;
@@ -280,5 +285,32 @@ public class LitmusPlayer implements
 
     @Override
     public void onDrmSessionManagerError(Exception e) {
+    }
+
+    public class WidevineTestMediaDrmCallback implements MediaDrmCallback {
+
+        private static final String WIDEVINE_GTS_DEFAULT_BASE_URI =
+                "http://wv-staging-proxy.appspot.com/proxy?provider=YouTube&video_id=";
+
+        private final String defaultUri;
+
+        public WidevineTestMediaDrmCallback(String videoId) {
+            defaultUri = WIDEVINE_GTS_DEFAULT_BASE_URI + videoId;
+        }
+
+        @Override
+        public byte[] executeProvisionRequest(UUID uuid, MediaDrm.ProvisionRequest request) throws IOException {
+            String url = request.getDefaultUrl() + "&signedRequest=" + new String(request.getData());
+            return Util.executePost(url, null, null);
+        }
+
+        @Override
+        public byte[] executeKeyRequest(UUID uuid, MediaDrm.KeyRequest request) throws IOException {
+            String url = request.getDefaultUrl();
+            if (TextUtils.isEmpty(url)) {
+                url = defaultUri;
+            }
+            return Util.executePost(url, request.getData(), null);
+        }
     }
 }
