--- client/android/WorldScope/app/src/main/java/com/litmus/worldscope/WidevineTestMediaDrmCallback.java
@@ -1,43 +1,0 @@
-package com.litmus.worldscope;
-
-import android.annotation.TargetApi;
-import android.media.MediaDrm;
-import android.text.TextUtils;
-
-import com.google.android.exoplayer.drm.MediaDrmCallback;
-import com.google.android.exoplayer.util.Util;
-
-import java.io.IOException;
-import java.util.UUID;
-
-/**
- * A {@link MediaDrmCallback} for Widevine test content.
- */
-@TargetApi(18)
-public class WidevineTestMediaDrmCallback implements MediaDrmCallback {
-
-    private static final String WIDEVINE_GTS_DEFAULT_BASE_URI =
-            "http://wv-staging-proxy.appspot.com/proxy?provider=YouTube&video_id=";
-
-    private final String defaultUri;
-
-    public WidevineTestMediaDrmCallback(String videoId) {
-        defaultUri = WIDEVINE_GTS_DEFAULT_BASE_URI + videoId;
-    }
-
-    @Override
-    public byte[] executeProvisionRequest(UUID uuid, MediaDrm.ProvisionRequest request) throws IOException {
-        String url = request.getDefaultUrl() + "&signedRequest=" + new String(request.getData());
-        return Util.executePost(url, null, null);
-    }
-
-    @Override
-    public byte[] executeKeyRequest(UUID uuid, MediaDrm.KeyRequest request) throws IOException {
-        String url = request.getDefaultUrl();
-        if (TextUtils.isEmpty(url)) {
-            url = defaultUri;
-        }
-        return Util.executePost(url, request.getData(), null);
-    }
-
-}
